Switch to uv package manager (#1)
* Switch to uv package manager

* backend: dockerfile and readme updated

* backend: readme updated

* docs: updated project urls

* Switch to uv application project template. Add pydantic settings for enviroment variables.

* add ruff

* lint and format backend

---------
