--- backend/.env.example
@@ -1,3 +1,0 @@
-PROJECT_NAME=NextJS-FastAPI App
-API_V1_STR=/api/v1
-SECRET_KEY=your-secret-key-here
