--- backend/CORS
@@ -1,1 +1,0 @@
-CORS_ORIGINS=["http://localhost:3000"]
