--- backend/PostgreSQL
@@ -1,5 +1,0 @@
-POSTGRES_SERVER=db
-POSTGRES_USER=postgres
-POSTGRES_PASSWORD=postgres
-POSTGRES_DB=app_db
-POSTGRES_PORT=5432
