+++ backend/README.md
@@ -1,24 +1,41 @@
 # FastAPI Backend
 
-
 ## Setup
+
+### Requirements
+
+Install [uv package and project manager](https://docs.astral.sh/uv/):
+
+- For Linux and MacOS:
+```bash
+curl -LsSf https://astral.sh/uv/install.sh | less
+```
+- For Windows:
+```bash
+powershell -c "irm https://astral.sh/uv/install.ps1 | more"
+```
+
 
 ### Local Development
 
-1. Create a virtual environment:
+1. Sync the backend dependencies with the environment:
 ```bash
-python -m venv venv
-source venv/bin/activate  # On Windows: .\venv\Scripts\activate
+cd backend
+uv sync
 ```
+- (Optional) Activate the Python environment: 
+  ```bash
+  source .venv/bin/activate  # On Windows: .\.venv\Scripts\activate
+  ```
 
-2. Install dependencies:
+2. Start the PostgreSQL database:
 ```bash
-pip install -r requirements.txt
+docker compose up db -d
 ```
 
 3. Run the development server:
 ```bash
-uvicorn app.main:app --reload
+uv run fastapi dev
 ```
 
 ### Docker Development
@@ -29,12 +46,27 @@ The backend service is containerized using Docker. You can run it using Docker C
 docker compose up backend
 ```
 
+## Project Structure
+
+```
+backend/
+├── config/
+├── models/
+├── routers/
+├── schemas/
+├── database.py
+├── Dockerfile
+├── main.py
+├── pyproject.toml
+├── README.md
+└── uv.lock
+```
+
 ## API Documentation
 
 Once the server is running, you can access:
 - Interactive API documentation: http://localhost:8000/docs
 - Alternative API documentation: http://localhost:8000/redoc
-
 
 ## API Endpoints
 
@@ -57,8 +89,12 @@ Once the server is running, you can access:
 
 ## Development Tools
 
+- **uv**: For Python package management and virtual environments
 - **Ruff**: For linting and code formatting
 - **Black**: For code formatting
 - **Pydantic**: For data validation
 - **SQLAlchemy**: For database ORM
-- **Alembic**: For database migrations (if needed)
+- **Alembic**: For database migrations
+- **FastAPI**: For API development
+- **Uvicorn**: For ASGI server
+- **Docker**: For containerization
