+++ backend/pyproject.toml
@@ -1,48 +1,31 @@
-[tool.poetry]
-name = "backend"
+[project]
+name = "coachAI-backend"
 version = "0.1.0"
-description = "FastAPI backend for Next.js web application"
-authors = ["Your Name <your.email@example.com>"]
-
-[tool.poetry.dependencies]
-python = "^3.9"
-fastapi = "^0.104.0"
-uvicorn = "^0.23.2"
-sqlalchemy = "^2.0.22"
-pydantic = "^2.4.2"
-pydantic-settings = "^2.0.3"
-psycopg2-binary = "^2.9.9"
-alembic = "^1.12.0"
-python-jose = "^3.3.0"
-passlib = "^1.7.4"
-python-multipart = "^0.0.6"
-email-validator = "^2.0.0"
-
-[tool.poetry.dev-dependencies]
-pytest = "^7.4.2"
-httpx = "^0.25.0"
-
-[tool.ruff]
-line-length = 88
-target-version = "py39"
-select = [
-    "E",   # pycodestyle errors
-    "F",   # pyflakes
-    "I",   # isort
-    "B",   # flake8-bugbear
-    "C4",  # flake8-comprehensions
-    "UP",  # pyupgrade
+description = "A smart, voice-based training app for HR professionals to master difficult conversations through guided preparation, realistic simulations, and actionable feedback"
+readme = "README.md"
+requires-python = ">=3.9,<4.0"
+dependencies = [
+    "fastapi[standard]>=0.112.1",
+    "sqlalchemy>=2.0.22",
+    "pydantic>=2.4.2",
+    "pydantic-settings>=2.0.3",
+    "psycopg2-binary>=2.9.9",
+    "alembic>=1.12.0",
+    "python-jose>=3.3.0",
+    "passlib>=1.7.4",
+    "python-multipart>=0.0.6",
+    "email-validator>=2.0.0",
+    "docker>=7.1.0",   
 ]
-ignore = []
 
-[tool.ruff.isort]
-known-first-party = ["app"]
+[tool.uv.workspace]
+members = ["backend"]
 
-[tool.black]
-line-length = 88
-target-version = ["py39"]
-include = '\.pyi?$'
+[dependency-groups]
+dev = [
+    "ruff>=0.11.9",
+]
 
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+lint = [
+  "ruff"
+]
