--- backend/requirements.txt
@@ -1,16 +1,0 @@
-fastapi==0.109.2
-uvicorn==0.27.1
-sqlalchemy==2.0.27
-pydantic==2.6.1
-pydantic-settings>=2.0.3
-psycopg2-binary==2.9.9
-alembic==1.13.1
-python-jose>=3.3.0
-passlib>=1.7.4
-python-multipart>=0.0.6
-email-validator>=2.0.0
-pytest>=7.4.2
-httpx>=0.25.0
-ruff==0.2.1
-black==24.1.1
-python-dotenv==1.0.1
