+++ backend/schemas/message_schema.py
@@ -1,15 +1,19 @@
-from pydantic import BaseModel
 from datetime import datetime
+
+from pydantic import BaseModel
+
 
 class MessageBase(BaseModel):
     content: str
 
+
 class MessageCreate(MessageBase):
     pass
+
 
 class Message(MessageBase):
     id: int
     created_at: datetime
 
     class Config:
-        from_attributes = True 
+        from_attributes = True
